refactor: 공통 Button 컴포넌트 생성
